+++ classes/shopping_cart/service_provider.php
@@ -17,7 +17,10 @@
 namespace local_shopping_cart\shopping_cart;
 
 use context_system;
+use local_shopping_cart\local\cartstore;
 use local_shopping_cart\local\entities\cartitem;
+use local_shopping_cart\shopping_cart_history;
+use moodle_exception;
 
 /**
  * Shopping_cart subsystem callback implementation for local_shopping_cart, for testing, does not have any use for production.
@@ -37,6 +40,17 @@ class service_provider implements \local_shopping_cart\local\callback\service_pr
      * @return array
      */
     public static function load_cartitem(string $area, int $itemid, int $userid = 0): array {
+
+        global $DB;
+
+        // We might need to add additional information to the area.
+        // Therefore, we split it here.
+        if (strpos($area, '-') !== false) {
+            list($area, $addinfo) = explode('-', $area);
+        } else {
+            // Handle the case where "-" doesn't exist in $area.
+            $addinfo = $area;
+        }
 
         switch ($area) {
             case 'bookingfee':
@@ -70,7 +84,7 @@ class service_provider implements \local_shopping_cart\local\callback\service_pr
                     );
                 }
 
-                $imageurl = new \moodle_url('/local/shopping_cart/pix/rebookingcredit.png');
+                $imageurl = new \moodle_url('/local/shopping_cart/pix/rebook.png');
                 $cartitem = new cartitem(
                     $itemid,
                     get_string('rebookingcredit', 'local_shopping_cart'),
@@ -87,6 +101,137 @@ class service_provider implements \local_shopping_cart\local\callback\service_pr
                     1, // Rebookingcredit cannot be deleted.
                 );
                 return ['cartitem' => $cartitem];
+            case 'rebookitem':
+                $record = $DB->get_record('local_shopping_cart_history', ['id' => $itemid]);
+
+                if ($record->componentname === 'local_shopping_cart') {
+                    return [];
+                }
+
+                $imageurl = new \moodle_url('/local/shopping_cart/pix/rebook.png');
+                $cartitem = new cartitem(
+                    $itemid,
+                    get_string('rebooking', 'local_shopping_cart') . ': ' . $record->itemname,
+                    - ((float)$record->price + (float)$record->tax),
+                    $record->currency ?? 'EUR',
+                    'local_shopping_cart',
+                    'rebookitem',
+                    '',  // No item description for rebookitem.
+                    $imageurl->out(),
+                    time(),
+                    0,
+                    0,
+                    'A',
+                    0, // Rebook items can be deleted again.
+                );
+                return ['cartitem' => $cartitem];
+            case 'installments':
+
+                // First step: We have the itemid which refers to the shopping cart history id.
+                // Now we have itemid etc. of the original item.
+                // Now, we need to understand if this is the first item in the cart.
+                // For installments, we first just have the id of the original item.
+
+                // For installments, we get the original history record.
+                $record = $DB->get_record('local_shopping_cart_history', ['id' => $itemid]);
+
+                if ($record->componentname === 'local_shopping_cart') {
+                    return [];
+                }
+
+                // Now, we check if a payment was already made.
+                // Or how many installments there are still to pay.
+                $jsonobject = json_decode($record->json);
+
+                // Now we run through all the payments and check...
+                // ... if they are open or already in the cart.
+
+                $cartstore = cartstore::instance($userid);
+                $itemincart = $cartstore->get_item(
+                    $record->componentname,
+                    $record->area,
+                    $record->itemid);
+
+                // Number of installments still to pay.
+                $stilltopay = (int)$jsonobject->installments->installments;
+
+                foreach ($jsonobject->installments->payments as $payment) {
+
+                    // We might have the payment item already in the cart.
+                    if (!empty($itemincart)) {
+
+                        // Check if the price of the item is correct.
+                        $remainder = $itemincart['price'] % $payment->price;
+                        if (!empty($remainder)) {
+                            throw new moodle_exception('wronginstallmentprice', 'local_shopping_cart');
+                        }
+
+                        // We only change the price if it is below the due amount.
+                        if ($itemincart['price'] <= ($payment->price * ($stilltopay - 1))) {
+                            $itemincart['price'] += $payment->price;
+                        }
+
+                        // Return the updated item.
+                        return ['cartitem' => new cartitem(
+                            $itemid,
+                            $itemincart['itemname'],
+                            $itemincart['price'],
+                            $itemincart['currency'],
+                            'local_shopping_cart',
+                            'installments-' . $addinfo,
+                            $itemincart['description'],
+                            $itemincart['imageurl'],
+                            $itemincart['canceluntil'],
+                            $itemincart['serviceperiodstart'],
+                            $itemincart['serviceperiodend'],
+                            $itemincart['taxcategory'],
+                            $itemincart['nodelete'],
+                            $itemincart['costcenter'],
+                        )];
+                    }
+
+                    // Now check if this particular installement is already paid in DB.
+
+                    // If not, we just add this payment and abort the loop.
+                    $imageurl = new \moodle_url('/local/shopping_cart/pix/coins.png');
+                    $cartitem = new cartitem(
+                        $itemid,
+                        get_string('installment', 'local_shopping_cart'). ': ' . $record->itemname,
+                        $payment->price,
+                        $record->currency,
+                        'local_shopping_cart',
+                        'installments-' . $addinfo,
+                        get_string('installment', 'local_shopping_cart'),
+                        '',
+                        $record->canceluntil,
+                        $record->serviceperiodstart,
+                        $record->serviceperiodend,
+                        $record->taxcategory,
+                        0,
+                        $record->costcenter
+                    );
+                    return ['cartitem' => $cartitem];
+
+                }
+            case 'rebookingfee':
+                $imageurl = new \moodle_url('/local/shopping_cart/pix/coins.png');
+                $cartitem = new cartitem(
+                    $itemid,
+                    get_string('rebookingfee', 'local_shopping_cart'),
+                    get_config('local_shopping_cart', 'rebookingfee'),
+                    get_config('local_shopping_cart', 'globalcurrency') ?? 'EUR',
+                    'local_shopping_cart',
+                    'rebookingfee',
+                    '',  // No item description for booking fee.
+                    $imageurl->out(), // Fee image.
+                    time(),
+                    0,
+                    0,
+                    'A',
+                    1, // Booking fee cannot be deleted.
+                );
+                return ['cartitem' => $cartitem];
+                break;
         }
 
         $now = time();
@@ -135,7 +280,7 @@ class service_provider implements \local_shopping_cart\local\callback\service_pr
 
     /**
      * Callback function that unloads a cart item and thus frees
-     * Used only in test.php for test purchases.
+     * Used only in demo.php for test purchases.
      *
      * @param string $area
      * @param int $itemid An identifier that is known to the plugin