+++ download_cash_report.php
@@ -19,6 +19,7 @@
  * @copyright 2023 Wunderbyte Gmbh <info@wunderbyte.at>
  * @author  Bernhard Fischer
  * @license https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @package local_shopping_cart
  */
 
 use local_wunderbyte_table\wunderbyte_table;
@@ -93,6 +94,20 @@ $table->define_columns([
     'annotation',
     'usermodified',
 ]);
+
+// In the shopping cart config settings, we can set a row limit for download.
+// This is needed, so download does not crash if the cash report is too big.
+$limitpart = '';
+if ($limit = get_config('local_shopping_cart', 'downloadcashreportlimit')) {
+    if (!empty($limit)) {
+        $limitpart = "LIMIT $limit";
+        $table->sql->from = str_replace(
+            ") s1",
+            " $limitpart ) s1", // We inject the LIMIT here.
+            $table->sql->from
+        );
+    }
+}
 
 // File name and sheet name.
 $fileandsheetname = "cash_report";